feat: move write blocks for kubernetes
